Avoid errors with the ClientBuilder service
The ClientBuilder was registered twice, once by ServiceLoader and once manually. The ServiceLoader version failed due to a missing no-args constructor, and further failed due to not being prototype scoped. This commit removes the unnecessary service, advertising the service loader capability from the correct service without registering a type.

The unnecessary RxInvokerProvider service is also removed, and simply registered with the Client.

Signed-off-by: Tim Ward <timothyjward@apache.org>