Updated lifecycle using the UI thread pattern
There are numerous threading difficulties with maintaining a whiteboard. The existing model would often issue unnecessary change count updates, and sometimes send an update before a resource was ready. The cleanup lifecycle was also somewhat inconsistent. Rather than trying to find and fix the threading bugs this commit substantially rewrites the JerseyServiceRuntime to:

* Do the service tracking and management itself, simplifying the provider plugins
* Use an "update thread" which is responsible for triggering changes in the whiteboard separately from the service binding thread
* Providers now create Jersey Container objects for each application path which are then managed by the JerseyService runtime

The overall lifecycle is now substantially simpler. I believe that there is also more code that could be deleted in a subsequent cleanup commit.

N.B. One other significant change is that the whiteboard updates are now asynchronous. Where previously a thread registering a whiteboard service could get away with immediately calling it, that thread must now wait for the change count to be updated.

Signed-off-by: Tim Ward <timothyjward@apache.org>