Avoid replacing the default Http Context
The Servlet Whiteboard provider uses custom ServletContextHelper instances to provide isolation between REST applications. This is a good thing, but when no base was defined the default application was creating a ServletContextHelper which shadowed the default ServletContextHelper. This, in turn, causes unexpected changes in behaviour for people using the Servlet Whiteboard. To be good citizens we should use the default context for the default application when that is possible.

Signed-off-by: Tim Ward <timothyjward@apache.org>