Refactor JerseyApplication
The JerseyApplication contains a lot of lifecycle code that could be eliminated if the application is made immutable. This commit delays the creation of the JerseyApplication until first access, but then marks the provider as locked. This fits with the new usage pattern where providers are cloned rather than modified.

Signed-off-by: Tim Ward <timothyjward@apache.org>