Ensure that implementation modules are resolvable
We want the implementation to be easily usable in external projects. This means that we shouldn't require additional Jersey or other implementation dependencies to be included in the external project. We can (and should) require that specification dependencies (e.g. SCR, Servlet Whiteboard, an SLF4J logging impl) are provided by the user. This commit fixes the pom dependencies so that these projects can be resolved.

This commit also fixes the SSE fragment version range, and avoids exporting an internal Jersey package, instead publishing a proper API type.

Signed-off-by: Tim Ward <timothyjward@apache.org>