Fix issues with generics in extension services
The whiteboard implementation must create dynamic classes for each extension service to allow multiple instances of the same class to be registered. The existing generics handling was limited, and could not correctly represent unbound type variables, wildcards, nested generics, or reification of variables declared in super types. This commit enhances the generics processing in an attempt to cover all cases. Fixes #39

Signed-off-by: Tim Ward <timothyjward@apache.org>